requirements.txt
alembic==1.6.2
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
importlib-metadata==4.0.1
Mako==1.1.4
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
pydantic==1.8.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
six==1.9.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
zipp==3.4.1
